+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - Sniff to disallow spaces when passing variables as reference
+
+### Updated
+- Code now adheres own codestyle
+
+## [13.0.1] - 2019-11-13
+## Fixed
+- Add squizlabs/php_codesniffer and sebastian/recursion-context as root dependencies in composer.json
 
 ## [13.0.0] - 2019-11-12
 ### Added